+++ raphodo/thumbnaildaemon.py
@@ -34,7 +34,8 @@ __copyright__ = "Copyright 2015-2016, Damon Lynch"
 import logging
 import pickle
 import sys
-from typing import Set
+import os
+from typing import Set, Dict
 
 from gettext import gettext as _
 import zmq
@@ -47,6 +48,7 @@ from raphodo.interprocess import (ThumbnailDaemonData, GenerateThumbnailsResults
                                   ThumbnailExtractorArgument)
 from raphodo.rpdfile import RPDFile
 from raphodo.thumbnailpara import GetThumbnailFromCache, preprocess_thumbnail_from_disk
+from raphodo.cache import FdoCacheLarge, FdoCacheNormal
 
 
 class DameonThumbnailWorker(DaemonProcess):
@@ -81,6 +83,10 @@ class DameonThumbnailWorker(DaemonProcess):
         assert data.frontend_port is not None
         self.frontend.connect("tcp://localhost:{}".format(data.frontend_port))
 
+        # handle freedesktop.org cache files directly
+        fdo_cache_large = FdoCacheLarge()
+        fdo_cache_normal = FdoCacheNormal()
+
         while True:
             # rename file and move to generated subfolder
             directive, content = self.receiver.recv_multipart()
@@ -89,59 +95,94 @@ class DameonThumbnailWorker(DaemonProcess):
 
             data = pickle.loads(content) # type: ThumbnailDaemonData
             rpd_file = data.rpd_file
-            rpd_file.modified_via_daemon_process = True
-            try:
+            if data.backup_full_file_names is not None:
+                # File has been backed up, and an extractor has already generated a FDO thumbnail
+                # for it.
+                # Copy and modify the existing FDO thumbnail
 
-                # Check the download source to see if it's in the caches, not the file
-                # we've just downloaded
+                # MD5 name of the existing FDO thumbnail
+                md5_name = data.fdo_name
+                assert md5_name
 
-                use_thumbnail_cache = (data.use_thumbnail_cache and not
-                    (data.write_fdo_thumbnail and rpd_file.should_write_fdo()))
-                cache_search = thumbnail_caches.get_from_cache(
-                    rpd_file=rpd_file,
-                    use_thumbnail_cache=use_thumbnail_cache)
-                task, thumbnail_bytes, full_file_name_to_work_on, origin = cache_search
-                processing = set()  # type: Set[ExtractionProcessing]
+                for backup_full_file_name in data.backup_full_file_names:
 
-                if task == ExtractionTask.undetermined:
-                    # Thumbnail was not found in any cache: extract it
+                    # Check to see if existing thumbnail in FDO cache can be
+                    # modified and renamed to reflect new URI
+                    try:
+                        mtime = os.path.getmtime(backup_full_file_name)
+                    except OSError:
+                        logging.debug("Backup file does not exist: %s", backup_full_file_name)
+                    else:
+                        logging.debug("Copying and modifying existing FDO 128 thumbnail for %s",
+                                      backup_full_file_name)
+                        fdo_cache_normal.modify_existing_thumbnail_and_save_copy(
+                            existing_cache_thumbnail=md5_name,
+                            full_file_name=backup_full_file_name,
+                            size=rpd_file.size,
+                            modification_time=mtime)
 
-                    task = preprocess_thumbnail_from_disk(rpd_file=rpd_file,
-                                                          processing=processing)
-                    if task != ExtractionTask.bypass:
-                        if rpd_file.thm_full_name is not None:
-                            full_file_name_to_work_on = rpd_file.download_thm_full_name
-                        else:
-                            full_file_name_to_work_on = rpd_file.download_full_file_name
+                        logging.debug("Copying and modifying existing FDO 256 thumbnail for %s",
+                                      backup_full_file_name)
+                        fdo_cache_large.modify_existing_thumbnail_and_save_copy(
+                            existing_cache_thumbnail=md5_name,
+                            full_file_name=backup_full_file_name,
+                            size=rpd_file.size,
+                            modification_time=mtime)
+            else:
+                # file has just been downloaded and renamed
+                rpd_file.modified_via_daemon_process = True
+                try:
 
-                if task == ExtractionTask.bypass:
-                    self.content = pickle.dumps(GenerateThumbnailsResults(
-                        rpd_file=rpd_file, thumbnail_bytes=thumbnail_bytes),
-                        pickle.HIGHEST_PROTOCOL)
-                    self.send_message_to_sink()
+                    # Check the download source to see if it's in the caches, not the file
+                    # we've just downloaded
 
-                elif task != ExtractionTask.undetermined:
-                    # Send data to load balancer, which will send to one of its
-                    # workers
-
-                    self.content = pickle.dumps(ThumbnailExtractorArgument(
+                    use_thumbnail_cache = (data.use_thumbnail_cache and not
+                        (data.write_fdo_thumbnail and rpd_file.should_write_fdo()))
+                    cache_search = thumbnail_caches.get_from_cache(
                         rpd_file=rpd_file,
-                        task=task,
-                        processing=processing,
-                        full_file_name_to_work_on=full_file_name_to_work_on,
-                        secondary_full_file_name='',
-                        exif_buffer=None,
-                        thumbnail_bytes = thumbnail_bytes,
-                        use_thumbnail_cache=data.use_thumbnail_cache,
-                        file_to_work_on_is_temporary=False,
-                        write_fdo_thumbnail=data.write_fdo_thumbnail),
-                        pickle.HIGHEST_PROTOCOL)
-                    self.frontend.send_multipart([b'data', self.content])
-            except SystemExit as e:
-                sys.exit(e)
-            except:
-                logging.error("Exception working on file %s", rpd_file.full_file_name)
-                logging.exception("Traceback:")
+                        use_thumbnail_cache=use_thumbnail_cache)
+                    task, thumbnail_bytes, full_file_name_to_work_on, origin = cache_search
+                    processing = set()  # type: Set[ExtractionProcessing]
+
+                    if task == ExtractionTask.undetermined:
+                        # Thumbnail was not found in any cache: extract it
+
+                        task = preprocess_thumbnail_from_disk(rpd_file=rpd_file,
+                                                              processing=processing)
+                        if task != ExtractionTask.bypass:
+                            if rpd_file.thm_full_name is not None:
+                                full_file_name_to_work_on = rpd_file.download_thm_full_name
+                            else:
+                                full_file_name_to_work_on = rpd_file.download_full_file_name
+
+                    if task == ExtractionTask.bypass:
+                        self.content = pickle.dumps(GenerateThumbnailsResults(
+                            rpd_file=rpd_file, thumbnail_bytes=thumbnail_bytes),
+                            pickle.HIGHEST_PROTOCOL)
+                        self.send_message_to_sink()
+
+                    elif task != ExtractionTask.undetermined:
+                        # Send data to load balancer, which will send to one of its
+                        # workers
+
+                        self.content = pickle.dumps(ThumbnailExtractorArgument(
+                            rpd_file=rpd_file,
+                            task=task,
+                            processing=processing,
+                            full_file_name_to_work_on=full_file_name_to_work_on,
+                            secondary_full_file_name='',
+                            exif_buffer=None,
+                            thumbnail_bytes = thumbnail_bytes,
+                            use_thumbnail_cache=data.use_thumbnail_cache,
+                            file_to_work_on_is_temporary=False,
+                            write_fdo_thumbnail=data.write_fdo_thumbnail),
+                            pickle.HIGHEST_PROTOCOL)
+                        self.frontend.send_multipart([b'data', self.content])
+                except SystemExit as e:
+                    sys.exit(e)
+                except:
+                    logging.error("Exception working on file %s", rpd_file.full_file_name)
+                    logging.exception("Traceback:")
 
 if __name__ == '__main__':
     generate_thumbnails = DameonThumbnailWorker()